@@ -11,7 +11,7 @@ import 'path_list.dart';
 /// 页面层级关系: 主页面 > 含有图片/视频的文件夹路径 > 该指定文件夹中的图片/视频列表 > 点击查看图片/播放视频
 
 class LocalMedia extends StatefulWidget {
-  const LocalMedia({Key? key}) : super(key: key);
+  const LocalMedia({super.key});
 
   @override
   State<LocalMedia> createState() => _LocalMediaState();
