@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:photo_manager/photo_manager.dart';
+import 'package:photo_manager_image_provider/photo_manager_image_provider.dart';
 
 class ImageItemWidget extends StatelessWidget {
   const ImageItemWidget({
-    Key? key,
+    super.key,
     required this.entity,
     required this.option,
     this.onTap,
     this.isLongPress,
-  }) : super(key: key);
+  });
 
   final AssetEntity entity;
   final ThumbnailOption option;
